requirements.txt
asgiref 
backports.zoneinfo 
certifi 
charset-normalizer==3.0.1
cloudinary==1.32.0
Django 
django-cloudinary-storage==0.3.0
django-fontawesome-5==1.0.18
django-tinymce==3.5.0
idna==3.4
mysqlclient==2.1.1
Pillow==9.4.0
python-decouple==3.8
requests==2.28.2
six==1.16.0
sqlparse 
tzdata 
urllib3==1.26.14
wincertstore==0.2
